feat(header): finalise desktop navbar
Used the trick of adding a pseudo-element to the desktop links to
make them fill the entire div. It's working beautifully now. For some reason, Chrome was being weird with <a> tags containing
blocks (it inserted two <a>'s??), so that wasn't an option.

Other changes include replacing the transition on the desktop with
an animation - this prevents weird hover effects. Also, I
standardised the pattern of `opacity-60` by default and full
opacity on hover.